Changes required for Extracting protein sequence for genetic positions #642 pull request
