+++ biojava-genome/src/main/java/org/biojava/nbio/genome/util/ChromosomeMappingTools.java
@@ -1,16 +1,22 @@
 package org.biojava.nbio.genome.util;
 
-import com.google.common.collect.Range;
-
-import org.biojava.nbio.genome.parsers.genename.ChromPos;
-import org.biojava.nbio.genome.parsers.genename.GeneChromosomePosition;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-
+import java.io.File;
+import java.io.IOException;
 import java.io.StringWriter;
 import java.util.ArrayList;
 import java.util.List;
+
+import org.biojava.nbio.core.exceptions.CompoundNotFoundException;
+import org.biojava.nbio.core.sequence.DNASequence;
+import org.biojava.nbio.core.sequence.compound.NucleotideCompound;
+import org.biojava.nbio.core.sequence.template.SequenceView;
+import org.biojava.nbio.genome.parsers.genename.ChromPos;
+import org.biojava.nbio.genome.parsers.genename.GeneChromosomePosition;
+import org.biojava.nbio.genome.parsers.twobit.TwoBitParser;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.google.common.collect.Range;
 
 /**
  *  A class that can map chromosomal positions to mRNA (coding sequence) positions.
@@ -22,14 +28,13 @@ public class ChromosomeMappingTools {
 
     private static final Logger logger = LoggerFactory.getLogger(ChromosomeMappingTools.class);
 
-
     private static final String newline = System.getProperty("line.separator");
-
 
     public static final String CHROMOSOME = "CHROMOSOME";
     public static final String CDS = "CDS";
-
-
+    
+	private TwoBitParser parser;
+	
     /** Pretty print the details of a GeneChromosomePosition to a String
      *
      * @param chromosomePosition
@@ -38,11 +43,8 @@ public class ChromosomeMappingTools {
     public static String formatExonStructure(GeneChromosomePosition chromosomePosition ){
         if ( chromosomePosition.getOrientation() == '+')
             return formatExonStructureForward(chromosomePosition);
-
         return formatExonStructureReverse(chromosomePosition);
-
     }
-
 
     private static String formatExonStructureForward(GeneChromosomePosition chromPos) {
 
@@ -81,7 +83,6 @@ public class ChromosomeMappingTools {
                 s.append("     UTR         : " + (cdsEnd +1) + " - " + format(end));
                 s.append(newline);
 
-
             } else if (inCoding) {
                 // full exon is coding
                 codingLength += (end - start);
@@ -89,15 +90,12 @@ public class ChromosomeMappingTools {
                 s.append("     Exon        : ").append(format(start + 1)).append(" - ").append(format(end)).append(" | ").append(Integer.toString(end - start)).append(" | ").append(Integer.toString(codingLength)).append(" | ").append(Integer.toString(codingLength % 3));
                 s.append(newline);
             }
-
         }
         s.append("Coding Length: ");
         s.append((codingLength-3)+"");
         s.append(newline);
         return s.toString();
     }
-
-
 
     private static String showGenePosLink(GeneChromosomePosition chromPos, Integer pos ) {
 
@@ -116,7 +114,6 @@ public class ChromosomeMappingTools {
 
         return buf.toString();
     }
-
 
     private static String formatExonStructureReverse(GeneChromosomePosition chromPos) {
         StringWriter s = new StringWriter();
@@ -237,7 +234,6 @@ public class ChromosomeMappingTools {
         return codingLength;
     }
 
-
     /**
      * maps the position of a CDS nucleotide back to the genome
      *
@@ -276,8 +272,6 @@ public class ChromosomeMappingTools {
      * @return
      */
     private static String format(int chromosomePosition){
-
-
         return String.format("%,d", chromosomePosition);
     }
 
@@ -543,7 +537,6 @@ public class ChromosomeMappingTools {
         return new ChromPos(-1,-1);
     }
 
-
     /**
      * Get the length of the coding sequence
      *
@@ -693,8 +686,6 @@ public class ChromosomeMappingTools {
         return codingLength-3 ;
     }
 
-
-
     /** Extracts the exon boundaries in CDS coordinates. (needs to be divided by 3 to get AA positions)
      *
      * @param chromPos
@@ -702,12 +693,9 @@ public class ChromosomeMappingTools {
      */
     public static List<Range<Integer>> getCDSExonRanges(GeneChromosomePosition chromPos){
         if ( chromPos.getOrientation() == '+')
-
             return getCDSExonRangesForward(chromPos,CDS);
-
         return getCDSExonRangesReverse(chromPos,CDS);
     }
-
 
     /** Extracts the boundaries of the coding regions in chromosomal coordinates
      *
@@ -717,7 +705,6 @@ public class ChromosomeMappingTools {
     public static List<Range<Integer>> getChromosomalRangesForCDS(GeneChromosomePosition chromPos){
         if ( chromPos.getOrientation() == '+')
             return getCDSExonRangesForward(chromPos,CHROMOSOME);
-
         return getCDSExonRangesReverse(chromPos,CHROMOSOME);
     }
 
@@ -790,7 +777,7 @@ public class ChromosomeMappingTools {
                 }
 
 
-                Range r ;
+                Range<Integer> r ;
                 if ( responseType.equals(CDS))
                     r = Range.closed(0,codingLength);
                 else
@@ -801,7 +788,7 @@ public class ChromosomeMappingTools {
             } else if (start <= cdsStart && end >= cdsStart) {
                 inCoding = false;
 
-                Range r;
+                Range<Integer> r;
                 if ( responseType.equals(CDS))
                     r = Range.closed(codingLength,codingLength+(end-cdsStart));
                 else
@@ -821,7 +808,7 @@ public class ChromosomeMappingTools {
             } else if (inCoding) {
                 // full exon is coding
 
-                Range r;
+                Range<Integer> r;
                 if ( responseType.equals(CDS))
                     r = Range.closed(codingLength,codingLength+(end-start));
                 else
@@ -850,8 +837,6 @@ public class ChromosomeMappingTools {
         return data;
     }
 
-
-
     private static List<Range<Integer>> getCDSExonRangesForward(GeneChromosomePosition chromPos,
                                                                 String responseType) {
 
@@ -878,7 +863,7 @@ public class ChromosomeMappingTools {
 //
 
 
-                Range r;
+                Range<Integer> r;
                 if ( responseType.equals(CDS))
                     r = Range.closed(0,codingLength);
                 else
@@ -890,7 +875,7 @@ public class ChromosomeMappingTools {
                 inCoding = false;
 
 
-                Range r;
+                Range<Integer> r;
                 if ( responseType.equals(CDS))
                     r = Range.closed(codingLength,codingLength+(cdsEnd-start));
                 else
@@ -901,7 +886,7 @@ public class ChromosomeMappingTools {
             } else if (inCoding) {
                 // full exon is coding
 
-                Range r;
+                Range<Integer> r;
                 if ( responseType.equals(CDS))
                     r = Range.closed(codingLength,codingLength+(end-start));
                 else
@@ -916,7 +901,6 @@ public class ChromosomeMappingTools {
         return data;
     }
 //
-
     /**
      * I have a genomic coordinate, where is it on the mRNA
      *
@@ -964,36 +948,15 @@ public class ChromosomeMappingTools {
         
         logger.debug("looking for CDS position for " +format(chromPos));
         
-        // remove exons that are fully landed in UTRs
-        List<Integer> tmpS = new ArrayList<Integer>(exonStarts);
-        List<Integer> tmpE = new ArrayList<Integer>(exonEnds);
-        
-        int j=0;
-        for (int i = 0; i < tmpS.size(); i++) {
-        	if ( ( tmpE.get(i) < cdsStart) || ( tmpS.get(i) > cdsEnd) ) {
-        		exonStarts.remove(j);
-        		exonEnds.remove(j);
-        	}
-        	else {
-        		j++;
-        	}
-        }
-        
-        // remove untranslated regions from exons
-        int nExons = exonStarts.size();
-        exonStarts.remove(0);
-        exonStarts.add(0, cdsStart);
-        exonEnds.remove(nExons-1);
-        exonEnds.add(cdsEnd);
+        // map the genetic coordinates of coding region on a stretch of a reverse strand
+        List<Range<Integer>> cdsRegions = getCDSRegionsForward(exonStarts, exonEnds, cdsStart, cdsEnd);
         
         int codingLength = 0;
         int lengthExon = 0;
-        
-        // map the genetic coordinate on a stretch of a reverse strand
-        for (int i = 0; i < nExons; i++) {
+        for (Range<Integer> range : cdsRegions) {
         	
-		    int start = exonStarts.get(i);
-		    int end = exonEnds.get(i);
+		    int start = range.lowerEndpoint();
+		    int end = range.upperEndpoint();
 		    
 		    lengthExon = end - start;
 
@@ -1030,13 +993,69 @@ public class ChromosomeMappingTools {
         }
         
         logger.debug("looking for CDS position for " +format(chromPos));
+                
+        // map the genetic coordinate on a stretch of a reverse strand
+        List<Range<Integer>> cdsRegions = getCDSRegionsReverse(exonStarts, exonEnds, cdsStart, cdsEnd);
         
+        int codingLength = 0;
+        int lengthExon = 0;
+        for ( int i=cdsRegions.size()-1; i>=0; i-- ) {
+        	
+		    int start = cdsRegions.get(i).lowerEndpoint();
+		    int end = cdsRegions.get(i).upperEndpoint();
+		    
+		    lengthExon = end - start;
+		    // +1 offset to be a base 1
+		    if (start+1 <= chromPos && end >= chromPos ) {
+		    	return codingLength + (end-chromPos+1);
+		    }
+	        else { 
+	        	codingLength += lengthExon;
+	        }
+        }
+        return -1;
+    }
+    
+	/** 
+	 *  Reads a genome from a locally stored .2bit file.
+	 */
+	public void readGenome(File file) throws Exception {
+		this.parser = new TwoBitParser(file);
+	}
+	
+	/** Sets a chromosome for TwoBitParser.
+	 * 
+	 * @param chr The chromosome name (e.g. chr21)
+	 */
+	public void setChromosome(String chr) throws Exception {
+		parser.close();
+		String[] names = parser.getSequenceNames();
+		for(int i=0;i<names.length;i++) {
+			if ( names[i].equalsIgnoreCase(chr) ) {
+				parser.setCurrentSequence(names[i]);
+				break;
+			}
+		}
+	}
+	
+    /** Extracts the exons boundaries in CDS coordinates for genes living on the reverse DNA strand.
+     *
+     * @param exonStarts The list holding the genetic coordinates pointing to the start positions of the exons (including UTR regions)  
+     * @param exonEnds The list holding the genetic coordinates pointing to the end positions of the exons (including UTR regions)
+     * @param cdsStart The start position of a coding region
+     * @param cdsEnd The end position of a coding region
+     * 
+     * @return the list of genetic positions corresponding to the exons boundaries in CDS coordinates
+    */
+    public static List<Range<Integer>> getCDSRegionsReverse(List<Integer> exonStarts, List<Integer> exonEnds,
+            int cdsStart, int cdsEnd) {
+
         // remove exons that are fully landed in UTRs
         List<Integer> tmpS = new ArrayList<Integer>(exonStarts);
         List<Integer> tmpE = new ArrayList<Integer>(exonEnds);
         
         int j=0;
-        for (int i = tmpS.size() - 1; i >= 0; i--) {
+        for (int i = 0; i < tmpS.size(); i++) {
         	if ( ( tmpE.get(i) < cdsStart) || ( tmpS.get(i) > cdsEnd) ) {
         		exonStarts.remove(j);
         		exonEnds.remove(j);
@@ -1053,24 +1072,97 @@ public class ChromosomeMappingTools {
         exonEnds.remove(nExons-1);
         exonEnds.add(cdsEnd);
         
-        int codingLength = 0;
-        int lengthExon = 0;
-        
-        // map the genetic coordinate on a stretch of a reverse strand
-        for (int i = nExons - 1; i >= 0; i--) {
-        	
-		    int start = exonStarts.get(i);
-		    int end = exonEnds.get(i);
-		    
-		    lengthExon = end - start;
-		    // +1 offset to be a base 1
-		    if (start+1 <= chromPos && end >= chromPos ) {
-		    	return codingLength + (end-chromPos+1);
-		    }
-	        else { 
-	        	codingLength += lengthExon;
-	        }
+        List<Range<Integer>> cdsRegion = new ArrayList<Range<Integer>>();
+        for ( int i=0; i<nExons; i++ ) {
+        	Range<Integer> r = Range.closed(exonStarts.get(i), exonEnds.get(i));
+        	cdsRegion.add(r);
         }
-        return -1;
+		return cdsRegion;
     }
+    
+    /** Extracts the exons boundaries in CDS coordinates for genes living on the forward DNA strand.
+    *
+    * @param exonStarts The list holding the genetic coordinates pointing to the start positions of the exons (including UTR regions)  
+    * @param exonEnds The list holding the genetic coordinates pointing to the end positions of the exons (including UTR regions)
+    * @param cdsStart The start position of a coding region
+    * @param cdsEnd The end position of a coding region
+    * 
+    * @return the list of genetic positions corresponding to the exons boundaries in CDS coordinates
+   */
+    public static List<Range<Integer>> getCDSRegionsForward(List<Integer> exonStarts, List<Integer> exonEnds,
+            int cdsStart, int cdsEnd) {
+    	
+        // remove exons that are fully landed in UTRs
+        List<Integer> tmpS = new ArrayList<Integer>(exonStarts);
+        List<Integer> tmpE = new ArrayList<Integer>(exonEnds);
+        
+        int j=0;
+        for (int i = 0; i < tmpS.size(); i++) {
+        	if ( ( tmpE.get(i) < cdsStart) || ( tmpS.get(i) > cdsEnd) ) {
+        		exonStarts.remove(j);
+        		exonEnds.remove(j);
+        	}
+        	else {
+        		j++;
+        	}
+        }
+        
+        // remove untranslated regions from exons
+        int nExons = exonStarts.size();
+        exonStarts.remove(0);
+        exonStarts.add(0, cdsStart);
+        exonEnds.remove(nExons-1);
+        exonEnds.add(cdsEnd);
+    	
+        List<Range<Integer>> cdsRegion = new ArrayList<Range<Integer>>();
+        for ( int i=0; i<nExons; i++ ) {
+        	Range<Integer> r = Range.closed(exonStarts.get(i), exonEnds.get(i));
+        	cdsRegion.add(r);
+        }
+		return cdsRegion;
+    }
+    
+    /** Extracts the DNA sequence transcribed from the input genetic coordinates.
+    *
+    * @param gcp The container with chromosomal positions
+    * 
+    * @return the DNA sequence transcribed from the input genetic coordinates
+   */
+    public String getTranscriptSequence(GeneChromosomePosition gcp) throws IOException, CompoundNotFoundException {
+    	return getTranscriptSequence(gcp.getExonStarts(), gcp.getExonEnds(), gcp.getCdsStart(), gcp.getCdsEnd(), gcp.getOrientation());
+    }
+    
+    /** Extracts the DNA sequence transcribed from the input genetic coordinates.
+    *
+    * @param exonStarts The list holding the genetic coordinates pointing to the start positions of the exons (including UTR regions)  
+    * @param exonEnds The list holding the genetic coordinates pointing to the end positions of the exons (including UTR regions)
+    * @param cdsStart The start position of a coding region
+    * @param cdsEnd The end position of a coding region
+    * @param orientation The orientation of the strand where the gene is living
+    * 
+    * @return the DNA sequence transcribed from the input genetic coordinates
+    */
+	public String getTranscriptSequence(List<Integer> exonStarts, List<Integer> exonEnds, int codingStart, int codingEnd, Character orientation) throws IOException, CompoundNotFoundException {
+
+		List<Range<Integer>> cdsRegion;
+		if (orientation.equals("-")) {
+			cdsRegion = getCDSRegionsReverse(exonStarts, exonEnds, codingStart, codingEnd);
+		}
+		else {
+			cdsRegion = getCDSRegionsForward(exonStarts, exonEnds, codingStart, codingEnd);
+		}
+
+		String transcription = "";
+		for (Range<Integer> range : cdsRegion) {
+			int length = range.upperEndpoint() - range.lowerEndpoint();
+			transcription += parser.loadFragment(range.lowerEndpoint(), length);
+		}
+		if (orientation.equals("-")) {
+			transcription = new StringBuilder(transcription).reverse().toString();
+			DNASequence dna = new DNASequence(transcription);
+			SequenceView<NucleotideCompound> compliment = dna.getComplement();
+			transcription = compliment.getSequenceAsString();
+		}
+		return transcription;
+	}
 }
